pyproject.toml
[project]
name = "clawbots"
version = "0.1.0"
description = "AI Agents in Virtual Worlds"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.0",
    "httpx>=0.26.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
]

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]
